requirements.txt
scikit-learn
xgboost
jupyterlab
opencv-python
matplotlib
numpy
pandas
torch
torchvision
Pillow 
